+++ install/migrations/0044_add_snapshot_value_to_content_import_records_table.sql
@@ -1,0 +1,1 @@
+ALTER TABLE {prefix}content_import_records ADD COLUMN snapshot_value LONGTEXT NULL AFTER external_id;